specification de chaque utilsateur à ses propres données

+++ controllers/expense.js
@@ -2,6 +2,7 @@
 const expenseShema = require('../models/expense')
 const financeShema = require('../models/finace')
 const budgetSchema = require('../models/budget')
+
 const APIResponse = require('../utils/APIResponse')
 const { getTotalExpense, getTotalIncomes } = require("../utils/operations")
 
@@ -10,6 +11,13 @@ const getExpenses = async (req, res) => {
 
     try {
         const result = await expenseShema.find({username})
+        
+        if(result.length < 1){
+            const message = `no data find in your expenses`
+            const successResponse = APIResponse.success({}, message)
+            return res.status(204).json(successResponse.toJSON())
+        }
+
         const successResponse = APIResponse.success(result, '')
         res.status(200).json(successResponse.toJSON())
 
@@ -36,13 +44,17 @@ const createExpenses = async (req, res) => {
         const budget = await budgetSchema.findOne({ username })
         const totalExpense = await getTotalExpense(username)
         const totalIncome = await getTotalIncomes(username)
-        const solde = budget.montant + (totalIncome - totalExpense)
-        await financeShema.findOneAndUpdate({username} ,{ totalExpense, solde })
 
-        const message = `expense created solde and total expense updated`
+        const solde = budget.montant + (totalIncome - totalExpense)
+        const bdgetUpdate =  budget.montant - totalExpense 
+        await financeShema.findOneAndUpdate({username} ,{ totalExpense, solde })
+        await budgetSchema.findOneAndUpdate({username} ,{ montant: bdgetUpdate })
+
+        const message = `expense created, budget, solde and total expense updated`
         const successResponse = APIResponse.success({}, message)
         res.status(201).json(successResponse.toJSON())
     } catch (error) {
+        console.log(error)
         res.status(400).json(error)
     }
 }
@@ -60,12 +72,22 @@ const deleteExpenses = async (req, res) => {
     }
 
     try {
+        const findExpenses = await expenseShema.find({username})
+        
+        if(findExpenses.length > 0){
+            const message = `no data find in your expenses`
+            const successResponse = APIResponse.success({}, message)
+            return res.status(204).json(successResponse.toJSON())
+        }
+
         await expenseShema.findOneAndDelete({username})
         const budget = await budgetSchema.findOne({ username })
         const totalIncome = await getTotalIncomes(username)
         const totalExpense = await getTotalExpense(username)
         const solde = budget.montant + (totalIncome - totalExpense)
         await financeShema.findOneAndUpdate({username} ,{ totalExpense, solde })
+        await budgetSchema.findOneAndUpdate({username} ,{ montant: budget.montant - totalExpense })
+
         
         const message = `expense for id ${id} deleted solde and total expense updated `
         const successResponse = APIResponse.success({}, message)
