+++ controllers/incomes.js
@@ -1,15 +1,23 @@
 
 const incomeShema = require('../models/incomes')
 const financeShema = require('../models/finace')
-const { getTotalIncomes } = require("../utils/operations")
+const budgetSchema = require('../models/budget')
+
+const { getTotalIncomes, getTotalExpense } = require("../utils/operations")
 const APIResponse = require('../utils/APIResponse')
 
 const getIncomes = async (req, res) => {
 
     const { username } = req.user
-    
     try {
         const result = await incomeShema.find({username})
+
+        if(result.length < 1){
+            const message = `no data find in your incomes`
+            const successResponse = APIResponse.success({}, message)
+            return res.status(204).json(successResponse.toJSON())
+        }
+
         const successResponse = APIResponse.success(result, '')
         res.status(200).json(successResponse.toJSON())
 
@@ -36,14 +44,16 @@ const createIncomes = async (req, res) => {
         const budget = await budgetSchema.findOne({ username })
         const totalIncome = await getTotalIncomes(username)
         const totalExpense = await getTotalExpense(username)
+
         const solde = budget.montant + (totalIncome - totalExpense)
         await financeShema.findOneAndUpdate({username} ,{ totalIncome, solde })
 
         const message = `income created solde and total income updated`
         const successResponse = APIResponse.success({}, message)
-        res.status(201).json(successResponse.toJSON())
+       return res.status(201).json(successResponse.toJSON())
 
     } catch (error) {
+        console.log(error)
         res.status(400).json(error)
     }
 }
@@ -61,6 +71,14 @@ const deleteIncomes = async (req, res) => {
     }
 
     try {
+        const findIncomes = await incomeShema.find({username})
+        
+        if(findIncomes.length > 0){
+            const message = `no data find in your incomes`
+            const successResponse = APIResponse.success({}, message)
+            return res.status(204).json(successResponse.toJSON())
+        }
+
         await incomeShema.findOneAndDelete({username})
         const totalIncome = await getTotalIncomes(username)
         const totalExpense = await getTotalExpense(username)
